Test Cleanup (#953)
1. Refactored `test/common` into separate files
2. Moved the `EventCounterCallback` into `tests/common`, since it will be used by non-checkpointing tests
3. Replaced the `SimpleBatchPairModel` with the `SimpleModel` throughout tests
4. Removed duplicated and now-unused test fixtures from `tests/fixtures/dummy_fixtures.py`
5. Suppressing `NoSchedulerWarning` by default
6. Moved utility tests into `tests/utils`. This folder now contains only tests on `composer.utils`, not test helper functions.
7. Fixed `test_state_serialize` by ensuring that each state has a unique model object and that `optimizer.zero_grad` is called
8. Changed default number of `in_features` to 1.
9. Moved `assert_state_equivalent` to `tests/common`, since it is used throughout other tests too.
10. Rename `TestMetricsCallback` to `MetricsCallback` since it is not a PyTest test class, and thus should not start with `Test`.
