+++ docs/source/doctest_fixtures.py
@@ -63,7 +63,7 @@ _repo_root = os.path.dirname(_docs_dir)
 if sys.path[0] != _repo_root:
     sys.path.insert(0, _repo_root)
 
-from tests.fixtures.models import SimpleBatchPairModel
+from tests.common import SimpleModel
 
 # Change the cwd to be the tempfile, so we don't pollute the documentation source folder
 tmpdir = tempfile.TemporaryDirectory()
@@ -74,9 +74,9 @@ num_channels = 3
 num_classes = 10
 data_shape = (num_channels, 5, 5)
 
-Model = SimpleBatchPairModel
+Model = SimpleModel
 
-model = SimpleBatchPairModel(num_channels, num_classes)
+model = SimpleModel(num_channels, num_classes)
 
 optimizer = torch.optim.SGD(model.parameters(), lr=0.001)
