+++ tests/callbacks/test_grad_monitor.py
@@ -14,7 +14,6 @@ def _do_trainer_fit(composer_trainer_hparams: TrainerHparams, log_layers: bool =
     composer_trainer_hparams.callbacks.append(grad_monitor_hparams)
     composer_trainer_hparams.max_duration = "1ep"
 
-    composer_trainer_hparams.train_batch_size = 50
     trainer = composer_trainer_hparams.initialize_object()
     log_destination = MagicMock()
     log_destination = cast(LoggerDestination, log_destination)
