+++ tests/fixtures/dummy_fixtures.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
-from typing import Tuple, Type, Union
-from unittest.mock import Mock
+from typing import Tuple, Union
 
 import pytest
 import torch
@@ -11,13 +10,12 @@ from torch.optim import Optimizer
 from composer.core import DataSpec, Precision, State
 from composer.core.types import DataLoader, PyTorchScheduler
 from composer.datasets import DataLoaderHparams, DatasetHparams
-from composer.loggers import Logger
-from composer.models import ComposerClassifier, ModelHparams
+from composer.models import ModelHparams
 from composer.optim import AdamHparams, ExponentialSchedulerHparams
 from composer.trainer import TrainerHparams
 from composer.trainer.devices import CPUDeviceHparams
-from tests.fixtures.models import (SimpleBatchPairModel, SimpleConvModel, _SimpleBatchPairModelHparams,
-                                   _SimpleDatasetHparams, _SimplePILDatasetHparams)
+from composer.trainer.trainer_hparams import dataset_registry, model_registry
+from tests.common import RandomClassificationDatasetHparams, SimpleModel, SimpleModelHparams
 
 
 @pytest.fixture
@@ -27,7 +25,7 @@ def dummy_in_shape() -> Tuple[int, ...]:
 
 @pytest.fixture
 def dummy_num_classes() -> int:
-    return 3
+    return 2
 
 
 @pytest.fixture()
@@ -46,21 +44,21 @@ def dummy_train_n_samples() -> int:
 
 
 @pytest.fixture
-def dummy_model_hparams(
-        dummy_in_shape: Tuple[int, ...], dummy_num_classes: int,
-        SimpleBatchPairModelHparams: Type[_SimpleBatchPairModelHparams]) -> _SimpleBatchPairModelHparams:
-    return SimpleBatchPairModelHparams(num_channels=dummy_in_shape[0], num_classes=dummy_num_classes)
+def dummy_model_hparams(dummy_in_shape: Tuple[int, ...], dummy_num_classes: int) -> SimpleModelHparams:
+    model_registry['simple'] = SimpleModelHparams
+    return SimpleModelHparams(num_features=dummy_in_shape[0], num_classes=dummy_num_classes)
 
 
 @pytest.fixture
-def dummy_model(dummy_model_hparams: _SimpleBatchPairModelHparams) -> SimpleBatchPairModel:
-    return dummy_model_hparams.initialize_object()
+def dummy_model(dummy_in_shape: Tuple[int, ...], dummy_num_classes: int) -> SimpleModel:
+    return SimpleModel(num_features=dummy_in_shape[0], num_classes=dummy_num_classes)
 
 
 @pytest.fixture
-def dummy_train_dataset_hparams(dummy_model: SimpleBatchPairModel, dummy_in_shape: Tuple[int],
-                                SimpleDatasetHparams: Type[_SimpleDatasetHparams]) -> DatasetHparams:
-    return SimpleDatasetHparams(
+def dummy_train_dataset_hparams(dummy_model: SimpleModel, dummy_in_shape: Tuple[int]) -> DatasetHparams:
+    dataset_registry['random_classification'] = RandomClassificationDatasetHparams
+    assert dummy_model.num_classes is not None
+    return RandomClassificationDatasetHparams(
         use_synthetic=True,
         drop_last=True,
         shuffle=False,
@@ -70,21 +68,10 @@ def dummy_train_dataset_hparams(dummy_model: SimpleBatchPairModel, dummy_in_shap
 
 
 @pytest.fixture
-def dummy_train_pil_dataset_hparams(dummy_model: SimpleBatchPairModel, dummy_in_shape: Tuple[int],
-                                    SimplePILDatasetHparams: Type[_SimplePILDatasetHparams]) -> DatasetHparams:
-    return SimplePILDatasetHparams(
-        use_synthetic=True,
-        drop_last=True,
-        shuffle=False,
-        num_classes=dummy_model.num_classes,
-        data_shape=list(dummy_in_shape)[1:],
-    )
-
-
-@pytest.fixture
-def dummy_val_dataset_hparams(dummy_model: SimpleBatchPairModel, dummy_in_shape: Tuple[int],
-                              SimpleDatasetHparams: Type[_SimpleDatasetHparams]) -> DatasetHparams:
-    return SimpleDatasetHparams(
+def dummy_val_dataset_hparams(dummy_model: SimpleModel, dummy_in_shape: Tuple[int]) -> DatasetHparams:
+    dataset_registry['random_classification'] = RandomClassificationDatasetHparams
+    assert dummy_model.num_classes is not None
+    return RandomClassificationDatasetHparams(
         use_synthetic=True,
         drop_last=False,
         shuffle=False,
@@ -94,7 +81,7 @@ def dummy_val_dataset_hparams(dummy_model: SimpleBatchPairModel, dummy_in_shape:
 
 
 @pytest.fixture
-def dummy_optimizer(dummy_model: SimpleBatchPairModel):
+def dummy_optimizer(dummy_model: SimpleModel):
     return torch.optim.SGD(dummy_model.parameters(), lr=0.001)
 
 
@@ -104,7 +91,7 @@ def dummy_scheduler(dummy_optimizer: Optimizer):
 
 
 @pytest.fixture()
-def dummy_state(dummy_model: SimpleBatchPairModel, dummy_train_dataloader: DataLoader, dummy_optimizer: Optimizer,
+def dummy_state(dummy_model: SimpleModel, dummy_train_dataloader: DataLoader, dummy_optimizer: Optimizer,
                 dummy_scheduler: PyTorchScheduler, rank_zero_seed: int) -> State:
     state = State(
         model=dummy_model,
@@ -138,42 +125,9 @@ def dummy_train_dataloader(dummy_train_dataset_hparams: DatasetHparams, dummy_tr
 
 
 @pytest.fixture
-def dummy_train_pil_dataloader(dummy_train_pil_dataset_hparams: DatasetHparams, dummy_train_batch_size: int,
-                               dummy_dataloader_hparams: DataLoaderHparams) -> Union[DataLoader, DataSpec]:
-    return dummy_train_pil_dataset_hparams.initialize_object(dummy_train_batch_size, dummy_dataloader_hparams)
-
-
-@pytest.fixture
 def dummy_val_dataloader(dummy_train_dataset_hparams: DatasetHparams, dummy_val_batch_size: int,
                          dummy_dataloader_hparams: DataLoaderHparams) -> Union[DataLoader, DataSpec]:
     return dummy_train_dataset_hparams.initialize_object(dummy_val_batch_size, dummy_dataloader_hparams)
-
-
-@pytest.fixture()
-def dummy_logger(dummy_state: State):
-    return Logger(dummy_state)
-
-
-"""
-Dummy algorithms
-"""
-
-
-@pytest.fixture()
-def algorithms(always_match_algorithms):
-    return always_match_algorithms
-
-
-@pytest.fixture()
-def always_match_algorithms():
-    attrs = {'match.return_value': True}
-    return [Mock(**attrs) for _ in range(5)]
-
-
-@pytest.fixture()
-def never_match_algorithms():
-    attrs = {'match.return_value': False}
-    return [Mock(**attrs) for _ in range(5)]
 
 
 @pytest.fixture
@@ -210,44 +164,3 @@ def composer_trainer_hparams(
         train_subset_num_batches=3,
         eval_subset_num_batches=3,
     )
-
-
-@pytest.fixture()
-def simple_conv_model_input():
-    return torch.rand((64, 32, 64, 64))
-
-
-@pytest.fixture()
-def state_with_model(simple_conv_model: torch.nn.Module, dummy_train_dataloader: DataLoader, rank_zero_seed: int):
-    return State(
-        grad_accum=1,
-        rank_zero_seed=rank_zero_seed,
-        max_duration="100ep",
-        model=simple_conv_model,
-        precision=Precision.FP32,
-        dataloader=dummy_train_dataloader,
-        dataloader_label="train",
-    )
-
-
-@pytest.fixture()
-def simple_conv_model():
-    return ComposerClassifier(SimpleConvModel())
-
-
-@pytest.fixture(scope="session")
-def SimpleBatchPairModelHparams():
-    TrainerHparams.register_class("model", _SimpleBatchPairModelHparams, "simple_batch_pair_model")
-    return _SimpleBatchPairModelHparams
-
-
-@pytest.fixture(scope="session")
-def SimpleDatasetHparams():
-    TrainerHparams.register_class("train_dataset", _SimpleDatasetHparams, "simple_dataset")
-    return _SimpleDatasetHparams
-
-
-@pytest.fixture(scope="session")
-def SimplePILDatasetHparams():
-    TrainerHparams.register_class("train_dataset", _SimplePILDatasetHparams, "simple_pil_dataset")
-    return _SimplePILDatasetHparams
