+++ tests/test_state.py
@@ -2,7 +2,6 @@
 
 import pathlib
 import random
-from typing import Any, Dict, List
 
 import torch
 import torch.nn.functional as F
@@ -10,20 +9,19 @@ from torch.functional import Tensor
 
 from composer.algorithms import ChannelsLastHparams
 from composer.core import DataSpec, Precision, State
-from composer.core.types import Batch, DataLoader
+from composer.core.types import Batch
 from composer.datasets.dataloader import DataLoaderHparams
 from composer.datasets.hparams import DatasetHparams
-from composer.models.base import ComposerModel
-from tests.fixtures.models import SimpleBatchPairModel
+from tests.common import SimpleModel, assert_state_equivalent
 
 
 def random_tensor(size=(4, 10)):
     return torch.rand(*size)
 
 
-def get_dummy_state(model: ComposerModel, train_dataloader: DataLoader):
+def get_dummy_state():
+    model = SimpleModel()
     optimizers = torch.optim.Adadelta(model.parameters())
-
     state = State(model=model,
                   grad_accum=random.randint(0, 100),
                   rank_zero_seed=random.randint(0, 100),
@@ -35,58 +33,15 @@ def get_dummy_state(model: ComposerModel, train_dataloader: DataLoader):
     state.loss = random_tensor()
     state.batch = (random_tensor(), random_tensor())
     state.outputs = random_tensor()
-    state.set_dataloader(train_dataloader, "train")
     return state
-
-
-def _check_item(item1: Any, item2: Any, path: str):
-    if item1 is None:
-        assert item2 is None, f"{path} differs: {item1} != {item2}"
-        return
-    if isinstance(item1, (str, float, int, bool)):
-        assert type(item1) == type(item2)
-        assert item1 == item2, f"{path} differs: {item1} != {item2}"
-        return
-    if isinstance(item1, torch.Tensor):
-        assert isinstance(item2, torch.Tensor)
-        # Using a high tolerance, as deepspeed non-determinisim can cause
-        # metric values to be off.
-        assert item1.allclose(item2, rtol=0.1, atol=0.1), f"{path} differs"
-        return
-    if isinstance(item1, dict):
-        assert isinstance(item2, dict), f"{path} differs: {item1} != {item2}"
-        _check_dict_recursively(item1, item2, path)
-        return
-    if isinstance(item1, list):
-        assert isinstance(item2, list), f"{path} differs: {item1} != {item2}"
-        _check_list_recursively(item1, item2, path)
-        return
-    raise NotImplementedError(f"Unsupported item type: {type(item1)}")
-
-
-def _check_list_recursively(list1: List[Any], list2: List[Any], path: str):
-    assert len(list1) == len(list2), f"{path} differs: {list1} != {list2}"
-    for i, (item1, item2) in enumerate(zip(list1, list2)):
-        _check_item(item1, item2, f"{path}/{i}")
-
-
-def _check_dict_recursively(dict1: Dict[str, Any], dict2: Dict[str, Any], path: str):
-    assert len(dict1) == len(dict2), f"{path} differs: {dict1} != {dict2}"
-    for k, val1 in dict1.items():
-        val2 = dict2[k]
-        _check_item(val1, val2, f"{path}/{k}")
-
-
-def assert_state_equivalent(state1: State, state2: State):
-    assert state1.serialized_attributes == state2.serialized_attributes
-    assert state1.is_model_deepspeed == state2.is_model_deepspeed
-
-    _check_dict_recursively(state1.state_dict(), state2.state_dict(), "")
 
 
 def train_one_step(state: State, batch: Batch) -> None:
     _, y = batch
     state.batch = batch
+
+    for optimizer in state.optimizers:
+        optimizer.zero_grad()
 
     state.outputs = state.model(state.batch_pair)
     assert isinstance(y, Tensor)
@@ -111,16 +66,11 @@ def get_batch(dataset_hparams: DatasetHparams, dataloader_hparams: DataLoaderHpa
 
 def test_state_serialize(
     tmpdir: pathlib.Path,
-    dummy_model: ComposerModel,
     dummy_dataloader_hparams: DataLoaderHparams,
     dummy_train_dataset_hparams: DatasetHparams,
-    dummy_train_dataloader: DataLoader,
 ):
-
-    assert isinstance(dummy_model, SimpleBatchPairModel)
-
-    state1 = get_dummy_state(dummy_model, dummy_train_dataloader)
-    state2 = get_dummy_state(dummy_model, dummy_train_dataloader)
+    state1 = get_dummy_state()
+    state2 = get_dummy_state()
 
     # train one step to set the optimizer states
     batch = get_batch(dummy_train_dataset_hparams, dummy_dataloader_hparams)
