--- tests/utils/model.py
@@ -1,31 +1,0 @@
-# Copyright 2021 MosaicML. All Rights Reserved.
-
-from typing import Any
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-
-class SimpleModel(nn.Module):
-
-    def __init__(self):
-        super().__init__()
-
-        self.conv1 = nn.Conv2d(1, 16, (3, 3), padding=0)
-        self.conv2 = nn.Conv2d(16, 32, (3, 3), padding=0, stride=2)
-        self.bn = nn.BatchNorm2d(32)
-        self.fc1 = nn.Linear(32 * 16, 32)
-        self.fc2 = nn.Linear(32, 10)
-
-    def forward(self, x: Any):
-        out = self.conv1(x)
-        out = F.relu(out)
-        out = self.conv2(out)
-        out = self.bn(out)
-        out = F.relu(out)
-        out = F.adaptive_avg_pool2d(out, (4, 4))
-        out = torch.flatten(out, 1, -1)
-        out = self.fc1(out)
-        out = F.relu(out)
-        return self.fc2(out)
