+++ tests/utils/test_string_enum.py
